# api-service/swagger-spec.json
{"openapi":"3.0.0","paths":{"/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"description":"User information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}}},"responses":{"200":{"description":"User created!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"400":{"description":"User with the requested e-mail already exists!"}},"tags":["auth"]}},"/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"200":{"description":"User authenticated, token generated!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"Invalid Credentials!"}},"tags":["auth"],"security":[{"basic":[]}]}},"/stock":{"get":{"operationId":"StockController_getStock","parameters":[{"name":"q","required":true,"in":"query","description":"Stock Quote","example":"AAPL.US","schema":{"type":"string"}}],"responses":{"200":{"description":"Stock Response returned!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Stock"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Stock not found!"}},"tags":["stock"],"security":[{"bearer":[]}]}},"/history":{"get":{"operationId":"StockController_getHistoryStock","parameters":[],"responses":{"200":{"description":"Stock History Response returned!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockHistory"}}}}},"401":{"description":"Unauthorized"}},"tags":["stock"],"security":[{"bearer":[]}]}},"/stats":{"get":{"operationId":"StockController_getStats","parameters":[],"responses":{"200":{"description":"Stats Response returned!","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Stat"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden Resource, only the permited role can do it."}},"tags":["stock"],"security":[{"bearer":[]}]}},"/reset-password":{"post":{"operationId":"SecurityController_requestResetPassword","parameters":[{"name":"email","required":true,"in":"query","description":"User Email","example":"jobsity@email.com","schema":{"type":"string"}}],"responses":{"200":{"description":"Password reset requested!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomResponse"}}}},"401":{"description":"Unauthorized"},"404":{"description":"User not found!"}},"tags":["security"],"security":[{"bearer":[]}]}},"/reset-password/confirm":{"post":{"operationId":"SecurityController_confirmResetPassword","parameters":[],"requestBody":{"required":true,"description":"User information to Confirm Reset Password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmResetPasswordDTO"}}}},"responses":{"200":{"description":"Password reseted!","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetedResponse"}}}},"401":{"description":"Unauthorized"},"404":{"description":"User not found!"},"406":{"description":"Invalid security code"}},"tags":["security"],"security":[{"bearer":[]}]}}},"info":{"title":"JobSity Challenge","description":"JobSity Challenge - Maick Speck","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"basic":{"type":"http","scheme":"basic"},"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDTO":{"type":"object","properties":{"email":{"type":"string","description":"User email","example":"jobsity@email.com"},"role":{"type":"string","description":"User role","example":"user","examples":["admin","user"]}},"required":["email","role"]},"UserResponse":{"type":"object","properties":{"email":{"type":"string","description":"User email","example":"jobsity@email.com"},"password":{"type":"string","description":"User password","example":"any_password"}},"required":["email","password"]},"TokenResponse":{"type":"object","properties":{"token":{"type":"string","description":"JWT Token","example":"any_jwt_token"}},"required":["token"]},"Stock":{"type":"object","properties":{"symbol":{"type":"string","description":"Stock Quote Symbol","example":"GOOGL.US"},"name":{"type":"string","description":"Stock Quote Name","example":"ALPHABET"},"date":{"type":"string","description":"Date Stock Quote","example":"2022-09-16"},"time":{"type":"string","description":"Date Time Stock Quote","example":"22:00:07"},"open":{"type":"number","description":"Open/First day value","example":102.07},"high":{"type":"number","description":"Highest day value","example":111.07},"low":{"type":"number","description":"Lowest day value","example":101.07},"close":{"type":"number","description":"Close/Last day value","example":105.07},"volume":{"type":"number","description":"Financial Transactions Volume","example":42797172}},"required":["symbol","name","date","time","open","high","low","close","volume"]},"StockHistory":{"type":"object","properties":{"date":{"type":"string","description":"Date Stock Quote","example":"2022-09-16"},"name":{"type":"string","description":"Stock Quote Name","example":"ALPHABET"},"symbol":{"type":"string","description":"Stock Quote Symbol","example":"GOOGL.US"},"open":{"type":"number","description":"Open/First day value","example":102.07},"high":{"type":"number","description":"Highest day value","example":111.07},"low":{"type":"number","description":"Lowest day value","example":101.07},"close":{"type":"number","description":"Close/Last day value","example":105.07}},"required":["date","name","symbol","open","high","low","close"]},"Stat":{"type":"object","properties":{"stock":{"type":"string","description":"Stock Quote Name","example":"ALPHABET"},"times_requested":{"type":"number","description":"Number of Requested times","example":25}},"required":["stock","times_requested"]},"CustomResponse":{"type":"object","properties":{"success":{"type":"boolean","description":"Success status","example":true},"response":{"type":"object","description":"Data response!","example":"Action requested!"}},"required":["success","response"]},"ConfirmResetPasswordDTO":{"type":"object","properties":{"email":{"type":"string","description":"User email","example":"jobsity@email.com"},"code":{"type":"number","description":"Security code to reset password","example":123456}},"required":["email","code"]},"PasswordResetedResponse":{"type":"object","properties":{"success":{"type":"boolean","description":"Success status","example":true},"response":{"type":"object","description":"Data response!","example":"Action requested!"},"password":{"type":"string","description":"New Generated Password","example":"new_password"}},"required":["success","response","password"]}}}}